Added Web Interface and Merged Scripts
